+++ tests/e2e/core-tests/specs/index.js
@@ -25,6 +25,8 @@ const runTaxSettingsTest = require( './merchant/wp-admin-settings-tax.test' );
 const runOrderStatusFiltersTest = require( './merchant/wp-admin-order-status-filters.test' );
 const runOrderRefundTest = require( './merchant/wp-admin-order-refund.test' );
 const runOrderApplyCouponTest = require( './merchant/wp-admin-order-apply-coupon.test' );
+const runProductEditDetailsTest = require( './merchant/wp-admin-product-edit-details.test' );
+const runProductSearchTest = require( './merchant/wp-admin-product-search.test' );
 
 const runSetupOnboardingTests = () => {
 	runActivationTest();
@@ -52,6 +54,8 @@ const runMerchantTests = () => {
 	runOrderStatusFiltersTest();
 	runOrderRefundTest();
 	runOrderApplyCouponTest();
+	runProductEditDetailsTest();
+	runProductSearchTest();
 }
 
 module.exports = {
@@ -76,5 +80,7 @@ module.exports = {
 	runOrderStatusFiltersTest,
 	runOrderRefundTest,
 	runOrderApplyCouponTest,
+	runProductEditDetailsTest,
+	runProductSearchTest,
 	runMerchantTests,
 };